requirements.txt
flask==2.3.3
flask-cors==4.0.0
pdfplumber==0.9.0
pandas==2.0.3
openpyxl==3.1.2
gunicorn==21.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
